configury: add version script preprocessing
In order to support FI_DIRECT, there will be a need
for a provider to make its symbols visible.  Since
libfabric currently uses a version script to control
visibility of libfabric symbols (on systems that support
it) we need a way to pre-process this file so that
a provider doesn't have to directly add symbols in
to the version script.

Also switch to feeding the version script in to a
more general loader script format in the event that
it might be useful to add further control on the way
the libfabric shared library is built in the future.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>